ovpn3cli/session: Improve session_start() helper sigaction implementation
The start_session() function would allocate a struct sigaction on the heap
and never release it.  While this isn't necessarily a huge issue, since this
function and the initial purpose of it is to be used in a shortlived program,
it would still be cleaner code fixing it.

Instead allocate the needed buffer on the stack during the lifetime of this
function call.  This serves the same purpose and there is no need to add
logic freeing any heap allocated memory.

Signed-off-by: David Sommerseth <dazo@eurephia.org>