client: Remove not needed virtual declaration on an override
The NetCfgTunBuilder::tun_builder_dco_new_peer() function overrides
a method inherited from openvpn::TunBuilderBase.  This is pulled in
via openvpn::ClientAPI::OpenVPNClient, which NetCfgTunBuilder use
as the main inheritance class.

In core-client-netcfg.hpp, the tun_builder_dco_new_peer() was tagged
as both virtual and override.  The virtual keyword should not be
used in this case.

Signed-off-by: David Sommerseth <dazo@eurephia.org>