distro/systemd: Add a failsafe starting the same configuration more times
Wuthout this change, it is possible to start a new VPN session multiple
times.  In many cases, this is won't work as systemd will also avoid
starting a running unit file.  But there are some cases where the
openvpn3-systemd script may have exited while the VPN session is
still running.

This adds a check if the current user has a VPN session running
with the same configuration profile name, and raises a RuntimeError
if this is the case.

Signed-off-by: David Sommerseth <dazo@eurephia.org>