events/status: Refactor Events::Status::PrintMode handling
The current implementation used a lot of static_cast<uint8>() in
addition to storing the print mode as an uint8_t while the enum class
for PrintMode was already defined.

This refactors the code to use the Status::PrintMode type throughout
the code.  A unit test for changing the print mode has also been
added.

Signed-off-by: David Sommerseth <dazo@eurephia.org>