client: Refactor DBus::Connection::Ptr passing
In both openvpn3-service-client and openvpn3-service-backendstart, a
DBus::Connection::Ptr is passed around.  This could be simplified a
bit.

This also avoids storing a copy of the connection object in the
ClientService as well, since it was only used in the constructor.
And the constructor has access to the connection object via
DBus::Service::GetConnection().

A few places the DBus::Connection::Ptr could be std::move()d.

Signed-off-by: David Sommerseth <dazo@eurephia.org>