configmgr: Fix failing TransferOwnership() D-Bus method
An implementation error since the initial GDBus++ migration was
discovered in the net.openvpn.v3.configuration.TransferOwnership()
D-Bus method input argument parsing.

When attempting to extract the new user id for the new owner of
a configuration profile, the wrong value index was used which
contains a string instead of uid_t, which makes the extraction
fail.

Signed-off-by: David Sommerseth <dazo@eurephia.org>