configmgr: Improve logging of imported persistent configuration profiles
This moves the main logging to happen after the persistent configuration
profile has been imported, where more information about the imported
profile can be extracted for the log.

The initial logging has been kept, but changed to be debug logging
instead, indicating that the configuration parsing is about to happen.

Signed-off-by: David Sommerseth <dazo@eurephia.org>