netcfg: Catch errors better when calling GetUID, GetPID and GetSubscriptionOwner
The GetUID(), GetPID() and GetSubscriptionOwner() proxy calls may
fail if the process doing a method call towards net.openvpn.v3.netcfg
disconnects.

This situation was not caught properly and triggered a lot of error
messages which could be confusing.  This implements a better approach
to handle the situation when the caller disappears.

Signed-off-by: David Sommerseth <dazo@eurephia.org>