netcfg/resolved: Add Link::WaitForBackgroundTasks() method
This helper method adds a synchronisation tool to wait for
all the running background tasks to complete before collecting
error messages.

This is most useful in test programs, which typically runs
for a short time.  Without this type of synchronisation, the
program would try to retrieve error messages before the running
background tasks had a chance to complete.  This typically ended
with no errors reported, which was not necessarily correct.

The current implementation to wait for the background tasks to
complete is incredibly simplistic and not very performant.  It
was considered that when the codebase moves to C++20 in the future,
better approaches for handling this can be used at that point.

Signed-off-by: David Sommerseth <dazo@eurephia.org>