netcfg/resolved: Switch SetDefaultRoute() to use SetLinkDefaultRoute() D-Bus method
The systemd-resolved D-Bus API provides two methods to set the
"default route" flag for network interfaces.  The currently used
one requires a D-Bus "link" object (/org/freedesktop/resolve1/link/...
D-Bus path).  Unfortunately, it can take a bit of time before the
systemd-resolved service has created this object, which delays
the configuration requested by openvpn3-service-netcfg.

The other systemd-resolved API only requires the service root
object (/org/freedesktop/resolve1 D-Bus path), where it provides
a SetLinkDefaultRoute() D-Bus method.  However, this API requires
to use the if_index as the key to configure this setting.  This
approach seems to be available quicker for openvpn3-service-netcfg.

With some refactoring, calling this method can now also be done
asynchronously.  This also helps the openvpn3-service-netcfg
performance a bit.  The downside is that we don't know more instantly
if the call succeeded or not, as a check to see if the method is
available or not.

In this context, this is considered less important today - as the
more common systemd-resolved versions do have this support.  The
availability of this feature will be detected, but at a later point
and there might be some users will experience more failing calls to
SetLinkDefaultRoute() on older systemd-resolved versions before it
stops doing these D-Bus calls.

Signed-off-by: David Sommerseth <dazo@eurephia.org>