netcfg/resolved: Rework background_call_data implementation
The struct background_call_data implementation was based on a raw
pointer approach and used a std::weak_ptr to the DBus::Proxy:Client
object it shares with the main resolved::Link object.

This approach had some issues in situations where the
openvpn3-service-netcfg service are under stress with lots of
interface destructions and creations.  The resolved::Link object may
be removed before the background thread has had a chance to do its
job and the receiving an invalid weak DBus::Proxy::Client ponter.

A better approach is to use a proper std::shared_ptr both for the
struct background_call_data as well as DBus::Proxy::Client object
used inside the lambda function.

The running_tasks_count counter is now also decreased in case of
errors as well, otherwise it would only be decreasd on success.

Signed-off-by: David Sommerseth <dazo@eurephia.org>