dbus: Add GDBus++ support function - LookupObject()
This functionality will appear in a newer GDBus++ version, but
OpenVPN 3 Linux is not able to upgrade a newer version currently.

The functionality provided by GDBusPP::Proxy::Utils::LookupObject()
is a lightweight version DBus::Proxy::Utils::Query::CheckObject().

CheckObject() will do extra checks that the D-Bus object is really
existing and present.  But it will also require a D-Bus policy to
be able to access the org.freedesktop.DBus.Properties.GetAll method
on that object.

The LookupObject() approach will just fetch the introspection XML
document of the parent object and check if the requested object
is found there as a '<node/>' element.

Signed-off-by: David Sommerseth <dazo@eurephia.org>