netcfg/proxy: Replace g_variant_new() with glib2::Builder
Many places in the NetCfgProxy::Device, g_variant_new() was used to
prepare the GVariant object containing the arguments to the D-Bus
method which is to be called.

GDBus++ provides a simpler wrapper, glib2::Builder::Add, which
handles the C++ data types more directly.  This makes the code
a bit easier to read and more consistent.

Signed-off-by: David Sommerseth <dazo@eurephia.org>