netcfg/dco: Refactor the DCO device teardown
The current implementation depended on the destructor of the NetCfgDCO
to tear down the DCO interface.  In theory, this works fine.  But in
practice, this can end up in a deadlock in the net.openvpn.v3.netcfg
service if there are rapid CreateVirtualInterface() and Destroy()
calls queued up.

The NetCfg::Device::destroy() function will call
DBus::Object::Manager::RemoveObject() to remove itself from the D-Bus.
This results in the NetCfg::Device destructor being called, and this
destructor will again call RemoveObject() on the NetCfgDCO object.

Both of these RemoveObject() calls are synchronised via a mutex
inside the GDBus++ library.  This results in the RemoveObject()
call via destroy() will set the mutex lock_guard, so when the
NetCfg:Device destructor is called and calls RemoveObject() again
it will deadlock because the NetCfgDCO destructor is not able to
complete its destruction.  So these two RemoveObject() call waits
for each other to complete.

When an openvpn3-service-client then tries to call the
CreateVirtualInterface() method, it again ends up unresponsive -
because creating a new D-Bus object depends on the same mutex lock
the two prior RemoveObject() methods is fighting over.  This again
results in the openvpn3-service-client to fail.

The solution is to start tearing down the tunbuilder and removing
the NetCfgDO interface before calling the RemoveObject() on the
main NetCfg::Device object.  Then the device destruction has a
more structured path which avoids the deadlock.

Signed-off-by: David Sommerseth <dazo@eurephia.org>