log/journald: Add O3_SENDER_PID as log event meta data
The systemd-journald logs _PID as the process ID writing the log
event.  In OpenVPN 3 Linux context, that will always point back at
the openvpn3-service-log service.  There are no track records of
the PID of the OpenVPN 3 Linux service sending the log event.

This detail is nice to have when debugging where you might have a PID
value of a crashed process.  With O3_SENDER_PID it is easy to extract
the log events for just this particular process.

Signed-off-by: David Sommerseth <dazo@eurephia.org>