netcfg: Don't attach logging for net.openvpn.v3.netcfg.core
The net.openvpn.v3.netcfg.core interface for log events has not been
in use since the migration to GDBus++, which happened in v22_dev.
All the OpenVPN 3 Core Library log events are now handled via the
CoreLog API and it reuses the LogSender object for the main
openvpn3-service-netcfg service for sending log events.

Signed-off-by: David Sommerseth <dazo@eurephia.org>