sessionmgr: Block starting duplicated sessions
The OpenVPN 3 Session Manager has allowed multiple sessions to be
started in parallel.  This will in most cases result in chaos, where
the duplicated tunnels will fight over the same network configuration
and in most cases block each other as the end effect.

This adds a fail-safe where it checks if the D-Bus configuration
profile path has already been started for the current user.  This
allows different users to use a shared VPN profile, as these setups
may be more clever in the network routes they set up.  But it
will block a single user to start multiple VPN sessions using the
same VPN profile.

At this point, the check is only against the D-Bus configuration
path and not the VPN configuration profile name.  This is because
the configuration profile names are not guaranteed to be unique,
while the D-Bus object paths must be unique.

Signed-off-by: David Sommerseth <dazo@eurephia.org>