common/utils: Slight refactoring of version retrival functions
This renames get_version() to get_program_version(), to try to make
it a bit clearer what differentiates it from the other version
functions.

The get_package_version() is a new function which just returns the
constexpr package_version string.  The declaration of this string
has been moved into utils.cpp and the new function is used instead
to make rebuilds with just a new version string in build-version.hpp
more efficient.  Now only a few programs needs to be recompiled
completely.  It is generally enough to compile utils.cpp and link
all the binaries without recompiling the rest of the program.

All these functions have now gotten more proper Doxygen documentation
to make the use cases clearer.

Signed-off-by: David Sommerseth <dazo@eurephia.org>