tests: Remove log-listener2 test program
This program has not been migrated to GDBus++, and the approach it
uses for receiving log events will not work any more.  Now it is
required to explicitly request the net.openvpn.v3.log service to
forward log events to it.  This is what the log-listener.cpp
test program already does.

Since this program was now the only one using the stop_handler(),
this could also be cleaned up in utils.cpp.

Signed-off-by: David Sommerseth <dazo@eurephia.org>