client: Log the D-Bus path to the NetCfg object on device creation
The client logs was lacking details about the virtual network
interface the net.openvpn.v3.netcfg service manages for the
client process.  This makes it harder to see which virtual
interfaces are related to which client sessions in the logs.

Since this information is not something normally needed, the
log level is set at LogCategory::VERB2 (log-level 5).

Signed-off-by: David Sommerseth <dazo@eurephia.org>