python: Add openvpn3-desktop-session-watcher
This is a very simple stand-alone utility/helper script which
observes VPN sessions from they start through the whole life-cycle
of the session and providing graphical desktop notifications when
the status changes.

In the case a VPN session requires a web-based authentication, the
URL to the authentication website is made easily available via the
desktop notification.

Signed-off-by: David Sommerseth <dazo@eurephia.org>