client: Plug a file descriptor leak with virtual tun interfaces
The file descriptor provided by the net.openvpn.v3.netcfg service
to create a new virtual tun interface was never properly closed.
That means that each time a client processed called the
NetCfgProxy::Establish() method, a new file descriptor would be
created and not necessarily properly closed.

To resolve this, the returned file descriptor is preserved inside
the NetCfgTunBuilder object and will be explicitly closed when the
virtual tun interface is torn down.

Signed-off-by: David Sommerseth <dazo@eurephia.org>